==> 3_RSF.Rmd <==
---
title: "Habitat_Suitability_Model_10-13"
author: "Imogen Schwandner -2671952S"
date: '2022-06-08'
output: html_document
---


A habitat-suitability model based on wildebeest GPS tracking data from 2010-2013. Built in a logistic regression framework with a logit link, based on location fixes as occurrences and pseudo-absences. Pseudo absences will be randomly generated within a buffer of the maximum observed step length of any animal in the data set.

1. Load libraries

```{r message=FALSE, warning=FALSE}
library(rgeos)
library(rgdal)
library(sp)
library(dplyr)
library(tidyverse)
library(raster)
library(sf)
library(amt)
library(lme4)
```

2. Load wildebeest data and covariate layers and Pseudo-absence Points

```{r}
load("WBS_covariates.RData")
load("Points.RData") #pseudo-absences (created in "Pseudo_Absences-1013.Rmd")
```


3. Editing WBS data to keep only columns of interest for HBS model     (AID and datetime)

```{r}
# keep only columns AID and datetime
WBS@data <- WBS@data[, c(1,22)]
```


4. Extracting Covariates for Presence-Points/ Occurrences

4.1. Creating a raster stack


```{r}
# stack covariate layers
WBS_stack <- stack(anthro, D_river, D_roadT1, D_roadT5, D_wood, TWI)
```

4.2. Extract Covariates for WBS Occurrences

```{r}
# extract raster values
rasValue <- extract(WBS_stack, WBS)
```

4.3 Convert to data frame and join raster Values

```{r}
# convert to data frame
WBS_RSF <- as.data.frame(WBS)

# add raster values to data frame
WBS_RSF <- cbind(WBS_RSF, rasValue)
```

4.4. Create Presence/Absence (Occurrence/Pseudo-Absence) column
```{r}
# Create P/A column with value 1
WBS_RSF$PA <- 1
```

5.   Extract Covariates for Pseudo-Absences
5.1. Extract values from raster stack of covariates
```{r}
# extract covariates from stack
rasValue1 <- extract(WBS_stack, Points)
```

5.2. Convert to data frame and join raster Values

```{r}
# convert to data framne
Points_RSF <- as.data.frame(Points)

# add raster values to data frame
Points_RSF <- cbind(Points_RSF, rasValue1)
```

```{r}
# Create P/A column with value 1
Points_RSF$PA <- 0
```

6. Add pseudo-absences (and covariates) to WBS_RSF
(need to reorder and change some columns)

```{r}
#rename anthropogenic risk column
WBS_RSF <- WBS_RSF %>%  rename(reprojanth_risk_clp = anth_risk_clp)

#reorder columns to match Points_RSF
WBS_RSF <- WBS_RSF[, c(4, 1, 2, 3, 5, 6, 7, 8, 9, 10, 11)]

#combine WBS Presence/Occurrence and Pseudo Absence Points into one data frame
RSF <- rbind(WBS_RSF, Points_RSF)
```

7. Add dynamic variables (NDVI and DNDI) extracted in separate script ("ndviExtract.R")

```{r}
# load ndvi and dndvi values extracted in "ndviExtract.R" script
load("ndvi.Rdata")

# attach to final data frame with Presences/ Occurrences and Pseudo-Absences
RSF <- cbind(RSF, ndvi)

# keep only columns of interest
RSF <- RSF[,1:11]
```


8. Build habitat suitability model

8.1 Edit covariates

```{r}
# rename covariates
names(RSF) <- c("datetime", "AID", "anthroRisk", "Driver", "Drdt1", "Drdt5", "Dwood", "TWI", "PA", "indvi", "dndvi")

```

8.2 Convert to data frame, remove NA's, adjust variables

```{r}
# convert to data frame
RSFdf <- as.data.frame(RSF)

# remove NA's (n = 14.000 or 0.003 % of observations, only within-pseudo-absences. Due to a few buffers extending outside the extent of covariate raster layers. This was limited by the supplied anthropogenic risk layer digitised in Stabach (2015).)
RSFdf <- RSFdf[complete.cases(RSFdf),]

# ensure AID is a factor
RSFdf$AID <- as.factor(RSFdf$AID)
str(data$AID)
```

8.3. Run a GLMM
```{r}
# using a generalized linear mixed effects model with a random effect for animal ID (AID) and a logit link function, where presence or absence (PA) depend on the covariates: NDVI, DNDVI (delta NDVI), TWI (topographic wetness index), and distances (including quadratic effects) to wooded areas, rivers as well as primary and secondary roads. All variables scaled.

mod <- glmer(PA ~ scale(indvi) +
                  scale(dndvi) + 
                  scale(TWI) + 
                  scale(Dwood) + 
                  scale(Driver) + 
                  scale(Drdt1) + 
                  scale(Drdt5) + 
                  scale(anthroRisk) + 
                  I(scale(Dwood)^2) + 
                  I(scale(Driver)^2) + 
                  I(scale(Drdt1)^2) + 
                  I(scale(Drdt5)^2) + 
                  (1 |AID), 
                data = RSFdf, family = binomial)

# model summary
summary(mod)

```

8.4. Model selection ( this might take a long time)
```{r}
step(mod)

#keep all covariates, quadratic terms and random effect

# model selection is theoretically also possible using the dredge() function in the MuMIn package, however failed in several attempts here
```

8.5. Performance

```{r}
r.squaredGLMM(mod)

# or (if it does run)
dredge(mod, extra = "R^2") # rank = "AIC"
```

8.6. Save model for predictions
```{r}
# save(mod, "mod.RData")
```
