==> 5_Connectivity_Analysis.Rmd <==
---
title: "Connectivity_Analysis"
author: "Imogen Schwandner -2671952S"
date: '2022-08-16'
output: html_document
---

Connectivity analysis run in Circuitscape v.4.0 (circuitscape.org) developed by McRae et al. (2008) based on GPS collar wildebeest movement data from 2010-2013 in th Greater Mara.

McRae, B. H., Dickson, B. G., Keitt, T. H., & Shah, V. B. (2008). USING CIRCUIT THEORY TO MODEL CONNECTIVITY IN ECOLOGY, EVOLUTION, AND CONSERVATION. Ecology, 89(10), 2712–2724. https://doi.org/10.1890/07-1861.1

1. Load libraries

```{r}
library(sp)
library(rgdal)
library(raster)
```

2. Load and edit input layers

2.1.    Resistance surface
2.1.1.  Load predicted habitat suitability
```{r}
# load predicted habitat suitability surface
suitability <-  raster("pred_rast.tif")
#plot(suitability)
```

2.1.1. Convert to resistance
```{r}
# invert suitability for resistance
resistance <- ((suitability - 1)*-1)
#plot(resistance)

# convert to ascii file (for use in Circuitscape)
#writeRaster(resistance, "resistance.asc")
```

2.1.2.   Incorporate fencing

Final output (fenceplr.asc) provided on github, however landDX fencing data set is available here: (was simply too large to upload)
https://www.arcgis.com/home/search.html?q=landDx_export

```{r}
### 2.1.2.1. Load fencing data
# read in lanDx fencing data
fencing <- readOGR("LandDx_Fencing_UTM36S_WGS84.shp")

### 2.1.2.2. Crop to resistance layer extent
# turn extent of resistance layer into polygon
e <- extent(resistance)
p <- as(e, 'SpatialPolygons')
# assign crs
crs(p) <- resistance@crs

# crop by extent of resistance (using p polygon)
fence <- crop(fencing, p)

### 2.1.2.3. Rasterize fencing layer
# rasterize
fence <- rasterize(fence, resistance, field = fence$active)

# potentially necessary adjustments for next step (mosaic) to work (very slight differences in each)
# resistance@extent <- fence@extent
# resistance@ncols <- fence@ncols
# resistance@nrows <- fence@nrows

### 2.1.2.4. Overlay resistance with fencelines
# create raster where resistance is overlayed with fence resistance (currently =1) and the maximum value of resistance (max 1) is kept
fenceresistance01 <- mosaic(resistance, fence, fun=max)

# re-assign crs to new layer
fenceresistance01@crs <- CRS("+init=epsg:32736")

# save file with resistance 0 to 1 (creation takes a long time)
#writeRaster(fenceresistance01, file = "fenceresistance_0_1.asc")

### 2.1.2.5. Assign resiatcne value of 100 to fencing
# assign resistance value of 100
fenceresistance01[fenceresistance01 == 1] <- 100

#write to file
#writeRaster(fenceresistance01, "fenceresistance100.asc", overwrite = TRUE)

### 2.1.2.6. Add and overlay fence polygons (areas fully enclosed by fences) created in QGIS v.3.16
#read polygonized fences
fencepolygon <- readOGR("fence_polygons.shp")

# add to fence resistance raster (already contains fence lines to account for fence lines that don't build a polygon)
fenceresistance100 <- raster("fenceresistance100.asc")
fenceresistance100@crs <- CRS("+init=epsg:32736")

### 2.1.2.7. Rasterize fence polygons
# rasterize
fencepolygonr <- rasterize(fencepolygon, fenceresistance100, field = fencepolygon$active)

# assign resistance value of 100 for polygons
fencepolygonr[fencepolygonr == 1] <- 100

# re-project
fencepolygonr@crs <- CRS("+init=epsg:32736")

#write as .asc file
#writeRaster(fencepolygonr, file = "fencepolygonr.asc")

### 2.1.2.8. Create single raster with fence lines and polygons
# add fence polygons onto resistance surface with fence lines
fenceplr <- mosaic(fenceresistance100, fencepolygonr, fun = max)

#write raster
#writeRaster(fenceplr, file = "fenceplr.asc")
```

2.2. Focal areas 

Focal areas set to be 
I) the corner point of Mara North Conservancy, were the Mara river restricts wildebeest movement to the west 
II) the centroid of the Naboisho Conservancy
III) Maji Moto town which lies central in the Loita Plains

```{r}
# read shapefile with the focal nodes (created in QGIS v.3.16)
MCPNMM <- readOGR("MCP_N_MM.shp")

# recode id as numeric for rasterizing
MCPNMM$id <- as.numeric(MCPNMM$id)

# rasterize to extend and resolution of resistance surface
MCPNMMr <- rasterize(MCPNMM, resistance, field = "id", fun = 'last')

# convert to ascii file (for use in Circuitscape)
#writeRaster(MCPNMMr, file = "MCPNMMraster.asc")
```

3.    Run Circuitscape
3.1.  PRE-FENCING (HISTORIC) CONNECTIVITY LEVELS 2010-2013

```{r}
########################
### Run Circuitscape  ##  -- PRE-FENCING CONNECTIVITY LEVELS
########################    


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMraster.asc",
                                  "resistance.asc",
                                  "PREFENCE_MCP_N_MM_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```

3.2. FENCED CONNECTIVITY LEVELS 2022

```{r}
# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMraster.asc",
                                  "fenceplr.asc",
                                  "FENCE_polygon_MCP_N_MM_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)

```
