==> 6_Restoration_and_Cost.Rmd <==
---
title: "Restoration and Cost"
author: "Imogen Schwandner -2671952S"
date: '2022-08-17'
output: html_document
---

Restoration of connectivity for migratory wildebeest in the Greater Mara by simulated fence removal along historic corridors. Three corridors were outlined in QGIS v.3.16 in an area of interest north-eats of Naboisho conservancy. Both previous connectivity levels and the amount of fence removal required were taken into account. For each corridor widths of 500m, 1km, 2km and 3km were tested, all fences within boundaries removed and connectivity levels across the region modelled in Circuitscape v.4.0. Then different scenarios were compared to baseline current and historic levels of connectivity in the area of interest.

Fence rasters of different scenarios were created in QGIS v.3.16 for ease of visual validation and reduced computational time but could have easily been created in R as well. A quick overview as to how they were derived is described below.


1. Load libraries

```{r}
library(rgdal)
library(raster)
library(rgeos)
library(sf)
```

2.   Load required data
2.1. Focal areas layer

Already created in "Connectivity_Analysis.Rmd".

```{r}
# read shapefile with 3 focal nodes
# MCPNMM <- readOGR("MCP_N_MM.shp")
# 
# reconfigure id column
# MCPNMM$id <- as.numeric(MCPNMM$id)
# 
# rasterize
# MCPNMMr <- rasterize(MCPNMM, resistance, field = "id", fun = 'last')
# 
# write to file
# writeRaster(MCPNMMr, file = "MCPNMMr.asc", overwrite = TRUE)

```

2.2. Resistance layer

```{r}
# read in resistance raster
suitability <- raster("suitability.tif")

# invert to resistance
resistance <- ((suitability - 1)*-1)

```


##############################################################
########################  CORRIDOR 1 #########################
##############################################################


2.3.  Load fence rasters for each corridor scenario
2.3.2 Load fence rasters 

```{r}
# 250m buffersize = 500m corridor width
C1250 <- raster("./restoration/500/C1250plr.tif")
# 1 km width
C1500 <- raster("./restoration/500/C1500plr.tif")
# 2 km width
C11000 <- raster("./restoration/500/C11000plr.tif")
# 3 km width
C11500 <- raster("./restoration/500/C11500plr.tif")
```

2.3.4. Re-adjustments

Minimal differences between raster layers impair Circuitscape functioning. Hence adjust resistance raster and focal areas raster to exactly match fence rasters, as it can sometimes cause issues.

```{r}
resistance@extent <- C1250@extent
resistance@ncols <- C1250250@ncols
resistance@nrows <- C1250@nrows

# read in file if not created here
# MCPNMMr <- raster("MCPNMMR.asc")
MCPNMMr@nrows <- C1250@nrows
MCPNMMr@ncols <- C1250@ncols
MCPNMMr@extent <- C1250@extent
# also re-assign crs
MCPNMMr@crs <- CRS("+init=epsg:32736")
# save to new file
writeRaster(MCPNMMr, file = "MCPNMMr.asc", overwrite = TRUE)
```


2.3.5. Recode fence resistance as 100 and assign no-data value to non-fence pixels

```{r}
C1250[C1250== 1,] <- 100
C1250[C1500 == 0,] <- NA
C1500[C1500== 1,] <- 100
C1500[C1500 == 0,] <- NA
C11000[C11000== 1,] <- 100
C11000[C11000 == 0,] <- NA
C11500[C11500== 1,] <- 100
C11500[C11500 == 0,] <- NA
```

2.3.6. Incorporate fence rasters into resistance layer

```{r}
# create single layer with habitat resistance and fences overlayed (keep maximum resistance for each pixel)
fenceresistanceC1250 <- mosaic(C1250, resistance, fun = max)
fenceresistanceC500 <- mosaic(C1500, resistance, fun = max)
fenceresistanceC11000 <- mosaic(C11000, resistance, fun = max)
fenceresistanceC11500 <- mosaic(C11500, resistance, fun = max)

# save new fence-resistance rasters as .asc files for Circuitscape
writeRaster(fenceresistanceC1250, "fenceresistanceC1250.asc")
writeRaster(fenceresistanceC500, "fenceresistanceC1500.asc")
writeRaster(fenceresistanceC11000, "fenceresistanceC11000.asc")
writeRaster(fenceresistanceC11500, "fenceresistanceC11500.asc")
```

3. Run Circuitscape

########################### C1250 ###############################

########################
### Run Circuitscape  ##  -- 250m radius 500m width corridor
########################

```{r}
# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC1250.asc",
                                  "C1250_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)

```


########################### C1500 ###############################

########################
### Run Circuitscape  ##  -- 500m radius 1km width corridor
########################

```{r}
# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC1500.asc",
                                  "C1500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


######################### C11000 ################################

########################
### Run Circuitscape  ##  -- 1km radius 2km width corridor
########################

```{r}
# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC11000.asc",
                                  "C11000_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


############################### C11500 ##############################

########################
### Run Circuitscape  ##  -- 1.5km radius 3km width corridor
########################

```{r}
# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC11500.asc",
                                  "C11500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


##############################################################
########################  CORRIDOR 2 #########################
##############################################################

2.3.    Load fence rasters for each corridor scenario
2.3.2.  Load fence rasters

```{r}
C2250 <- raster("./restoration/500/C2250plr.tif")
C2500 <- raster("./restoration/500/C2500plr.tif")
C21000 <- raster("./restoration/500/C21000plr.tif")
C21500 <- raster("./restoration/500/C21500plr.tif")
```

2.3.4. Recode fence resistance as 100 and assign no-data value to non-fence pixels

```{r}
C2250[C2250== 1,] <- 100
C2250[C2500 == 0,] <- NA
C2500[C2500== 1,] <- 100
C2500[C2500 == 0,] <- NA
C21000[C21000== 1,] <- 100
C21000[C21000 == 0,] <- NA
C21500[C21500== 1,] <- 100
C21500[C21500 == 0,] <- NA
```

2.3.5. Incorporate fence rasters into resistance layer

```{r}
# create single layer with habitat resistance and fences overlayed (keep maximum resistance for each pixel)
fenceresistanceC2250 <- mosaic(C2250, resistance, fun = max)
fenceresistanceC2500 <- mosaic(C2500, resistance, fun = max)
fenceresistanceC21000 <- mosaic(C21000, resistance, fun = max)
fenceresistanceC21500 <- mosaic(C21500, resistance, fun = max)

# save new fence-resistance rasters as .asc files for Circuitscape
writeRaster(fenceresistanceC2250, "fenceresistanceC2250.asc")
writeRaster(fenceresistanceC2500, "fenceresistanceC2500.asc")
writeRaster(fenceresistanceC21000, "fenceresistanceC21000.asc")
writeRaster(fenceresistanceC21500, "fenceresistanceC21500.asc")
```

3. Run Circuitscape

########################### C2250 ###############################

```{r}
########################
### Run Circuitscape  ##  -- 250m radius 500m width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC2250.asc",
                                  "C2250_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


############################ C2500 #############################

```{r}
########################
### Run Circuitscape  ##  -- 500m radius 1km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC2500.asc",
                                  "C2500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


######################### C21000 ################################

```{r}
########################
### Run Circuitscape  ##  -- 1km radius 2km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC21000.asc",
                                  "C21000_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```


############################### C21500 ##############################

```{r}
########################
### Run Circuitscape  ##  -- 1.5km radius 3km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC21500.asc",
                                  "C21500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)

```


##############################################################
########################  CORRIDOR 3 #########################
##############################################################

2.3.    Load fence rasters for each corridor scenario
2.3.2.  Load fence rasters

```{r}
C3250 <- raster("./restoration/500/C3250plr.tif")
C3500 <- raster("./restoration/500/C3500plr.tif")
C31000 <- raster("./restoration/500/C31000plr.tif")
C31500 <- raster("./restoration/500/C31500plr.tif")
```

2.3.4. Recode fence resistance as 100 and assign no-data value to non-fence pixels

```{r}
C3250[C3250== 1,] <- 100
C3250[C3500 == 0,] <- NA
C3500[C3500== 1,] <- 100
C3500[C3500 == 0,] <- NA
C31000[C31000== 1,] <- 100
C31000[C31000 == 0,] <- NA
C31500[C31500== 1,] <- 100
C31500[C31500 == 0,] <- NA
```

2.3.5. Incorporate fence rasters into resistance layer

```{r}
# create single layer with habitat resistance and fences overlayed (keep maximum resistance for each pixel)
fenceresistanceC3250 <- mosaic(C3250, resistance, fun = max)
fenceresistanceC3500 <- mosaic(C3500, resistance, fun = max)
fenceresistanceC31000 <- mosaic(C31000, resistance, fun = max)
fenceresistanceC31500 <- mosaic(C31500, resistance, fun = max)

# save new fence-resistance rasters as .asc files for Circuitscape
writeRaster(fenceresistanceC3250, "fenceresistanceC3250.asc")
writeRaster(fenceresistanceC3500, "fenceresistanceC3500.asc")
writeRaster(fenceresistanceC31000, "fenceresistanceC31000.asc")
writeRaster(fenceresistanceC31500, "fenceresistanceC31500.asc")

```

3. Run Circuitscpe

########################### C3250 ###############################

```{r}
########################
### Run Circuitscape  ##  -- 250m radius 500m width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC3250.asc",
                                  "C3250_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)

```

############################ C3500 #############################

```{r}
########################
### Run Circuitscape  ##  -- 500m radius 1km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC3500.asc",
                                  "C3500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)

```


######################### C31000 ################################

```{r}
########################
### Run Circuitscape  ##  -- 1km radius 2km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC31000.asc",
                                  "C31000_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```

############################### C31500 ##############################

```{r}
########################
### Run Circuitscape  ##  -- 1.5km radius 3km width corridor
########################


# Make a place holder for the cs_run.exe path
CS_exe <- 'C:/"Program Files"/Circuitscape/cs_run.exe'

# Make an .ini file
CS_ini <- c("[circuitscape options]",            
            "data_type = raster",
            "scenario = pairwise",
            "set_focal_node_currents_to_zero = True",
            "connect_four_neighbors_only = TRUE",
            "write_cur_maps = True",
            "write_cum_cur_map_only = True",
            "log_transform_maps = False",
            paste(c("point_file =",
                    "habitat_file =",
                    "output_file ="),
                  paste(getwd(),c("MCPNMMr.asc",
                                  "fenceresistanceC31500.asc",
                                  "C31500_CS.out"),
                        sep="/")))

# Write it to your working directory
writeLines(CS_ini,"myini.ini")

# Make the CS run cmd
CS_run <- paste(CS_exe, paste(getwd(),"myini.ini",sep="/")) # Make the cmd

# Run the command
system(CS_run)
```

4.   Comparison of scenarios
4.1. Load restoration extent

```{r}
# read in restoration extent (rectangular area covering all three corridors)
extent <- readOGR("restoration_extent.shp")
```

4.2. Read in and crop current (2022) and historic connectivity maps

```{r}
# crop pre-fencing and fenced scenario to extent for comparisons
## read in files
prefence <- raster("PREFENCE_MCP_N_MM_CS_cum_curmap.asc")
fence <- raster("FENCE_polygon_MCP_N_MM_CS_cum_curmap.asc")

# crop pre-fencing and fencing connectivity (cumulative current) maps to extent
rprefence <- crop(prefence, extent)
rfence <- crop(fence, extent)
```

4.3. Calculate connectivity metric - sum of cumulative current through the area for current and historic scenrio

```{r}
# get sum of cumulative current in the restoration area with and without fencing before restoration
prefencecurrent <- cellStats(rprefence, stat = sum)
# 773.3755
fencecurrent <- cellStats(rfence, stat = sum)
# 599.7039
```

4.4. Read restoration connectivity maps and calculate restored connectivity metric

Requires creation of separate folder "restoration_currentmaps" within the working directory and moving current maps of all scenarios within.

```{r}
# read in files of all scenarios at once
files <- list.files(path = "./restoration_currentmaps", pattern = ".asc$")

# set working directory for lapply to work
setwd("./restoration_currentmaps")

#import all raster files in folder using lapply
rasters <- lapply(files, raster)

#set it back if necessary
#setwd("C:/Users/Owner/Desktop/MScQMBCE/DISS/Circuitscape")

#crop files to restoration extent
newrasters <-  lapply(rasters, crop, extent)

# calculate sum of cumulative current (connectivity metric) for each scenario
cumcurs <- lapply(newrasters, cellStats, stat = sum)
```

4.5. Create data frame to store the results and add additional calculations

```{r}
# create data frame with file name and cumulative current sum
currents <-  data.frame("origin" = files, "sum_current" = unlist(cumcurs))

# add column identifier - which corridor
currents$corridor <- c(rep(1,4), rep(2,4), rep(3,4))
# make factor
currents$corridor <- as.factor(currents$corridor)

# add total width of corridor
currents$totalwidth <- rep(c(2,3,0.5,1),3)

# add column with the total length of fence (in km) removed for creation of corridor (calculated in QGIS)
currents$kmfenceremoved <- c(57.82, 87.62, 15.28, 28.73, 94.56, 136.29, 24.33, 46.87, 92.24, 139.71, 20.41, 42.28)


```

4.6. Additional calculations for different connectivity improvement metrics

```{r}
# improvement in connectivity as the percentage of the difference (gap) between the current and historic scenario bridged by the intervention
currents$gapfilled <- currents$improvement/currents$diffprefencefence *100 #retained for plotting

# improvement as the difference in total current/connectivity for each scenario compared to current fully fenced scenario
#currents$improvement <- currents$sum_current - fencecurrent

# difference in connectivity to historic scenario
#currents$dhistoric <- prefencecurrent - currents$sum_current 

# improval in connectivity as a percentage of the fenced scenario
#currents$percentimproval <- currents$improvement/ fencecurrent *100

# the difference in connectivity between the current and historic scenario
#currents$diffprefencefence <- prefencecurrent - fencecurrent

# the amount of connectivity retained in the current fenced scenario as a percentage of the historic connectivity
#fencepercentofprefence <- fencecurrent/prefencecurrent

# the improved connectivity in restoration scenarios as a percentage of the historic connectivity levels
#currents$percentofhistoric <- currents$sum_current/currents$prefencecurrent *100
```

5. Estimation of intervention costs

The estimation of intervention costs is based on cost figures derived from Pardamat Conservation Area. According to Jackson Sasine, current conservation manager at PCA, landowners in this area are paid 10.000 KES per acre of land, to permanently remove fences. 

In order to work with these figures, parcels assumed to be enclosed (at least three fence lines - of those two at least parallel or angled toward another) by the above identified fences, or those lying fully within a larger fully enclosed parcel extending outside the corridor, were digitised (using a polygonising algorithm and additional manual digitisation) in QGIS v.3.16. In QGIS the nine corridor polygon shapes (three locations, three sizes each) were used to select parcels intersecting the corridors, cropping at the corridor outline, to only account for areas within the corridor. This might underestimate costs, given that parcels frequently extend further and landowners might request their entire parcel to be compensated. The area of the remaining parcels was then estimated in acres to allow estimation of the total fenced area to be targeted for fence removal for each corridor.

5.1. Add total corridor area (in acres) to currents data frame
```{r}
currents$Carea_acre <- c(3746.08, 5486.83, 1975.19, # C1 -2-3-1- km
                         6087.24, 8860.50, 3178.34, # C2
                         3018.99, 5098.26, 1269.37) # C3
```

5.2. Calculate cost in KES (10.000 KES per acre)

```{r}
currents$costKES <- currents$Carea_acre *10000
```

5.3. Calculate cost in USD (1 KES = 0.0075 USD - 05/04/2023)

```{r}
currents$costUSD <- currents$costKES *0.0075
```

6. Plotting
```{r}
ggplot(currents, aes(fill=Corridor, y=gapfilled,
                     x=costUSD,
                     label = factor(totalwidth))) + 
  geom_bar(position="dodge", stat="identity") + # legend looks better than with geom_label
  geom_label(check_overlap = T, show_guide = F) +
  scale_fill_discrete(labels = c("I", "II", "III")) +
  scale_fill_manual(values=c("#3182bd", "#deebf7", "#9ecae1")) +
  labs( y = "Connectivity improvement
        between current and historic levels (in %)",
        x = "Cost of fence removal in USD") +
  scale_x_continuous(labels = scales::label_comma()) +
  scale_y_continuous(limits = c(35,55)) +
  theme_bw()

```
